+++ Sudo/CHANGELOG.md
@@ -2,10 +2,14 @@
 
 All changes to the Seamless Sudo utility will be documented in this file.
 
+## 1.1.1
+
+- Added the ability to run commands via Command Prompt (cmd.exe) or PowerShell (pwsh.exe/powershell.exe)
+
 ## 1.1.0
 
 - Integrated the Sudo daemon (Sud.dll) into the main executable module
-- Added the ability to run as SYSTEM by cloning the access token of winlogon.exe
+- Added the ability to run commands as SYSTEM by cloning the access token of winlogon.exe
 
 ## 1.0.0